fix(Composite Device): refactors target device behavior into its own struct
